+++ spec/factories/items.rb
@@ -1,0 +1,5 @@
+FactoryBot.define do
+  factory :item do
+    
+  end
+end
